requirements.txt
streamlit==1.39.0
PyYAML==6.0.2
st-gsheets-connection==0.1.0
ruff==0.6.9
